96252: Move Item page components out of SharedModule
