+++ src/app/item-page/alerts/item-alerts.component.spec.ts
@@ -2,7 +2,7 @@ import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 import { ItemAlertsComponent } from './item-alerts.component';
 import { TranslateModule } from '@ngx-translate/core';
 import { NO_ERRORS_SCHEMA } from '@angular/core';
-import { Item } from '../../../core/shared/item.model';
+import { Item } from '../../core/shared/item.model';
 import { By } from '@angular/platform-browser';
 
 describe('ItemAlertsComponent', () => {
