+++ src/app/item-page/bitstreams/request-a-copy/bitstream-request-a-copy-page.component.spec.ts
@@ -1,30 +1,30 @@
 import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
-import { AuthService } from '../../core/auth/auth.service';
+import { AuthService } from '../../../core/auth/auth.service';
 import { of as observableOf } from 'rxjs';
-import { Bitstream } from '../../core/shared/bitstream.model';
-import { AuthorizationDataService } from '../../core/data/feature-authorization/authorization-data.service';
+import { Bitstream } from '../../../core/shared/bitstream.model';
+import { AuthorizationDataService } from '../../../core/data/feature-authorization/authorization-data.service';
 import {
   createFailedRemoteDataObject$,
   createSuccessfulRemoteDataObject,
   createSuccessfulRemoteDataObject$
-} from '../remote-data.utils';
+} from '../../../shared/remote-data.utils';
 import { ActivatedRoute, Router } from '@angular/router';
 import { TranslateModule } from '@ngx-translate/core';
 import { CommonModule } from '@angular/common';
 import { BitstreamRequestACopyPageComponent } from './bitstream-request-a-copy-page.component';
 import { By } from '@angular/platform-browser';
-import { RouterStub } from '../testing/router.stub';
+import { RouterStub } from '../../../shared/testing/router.stub';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
-import { NotificationsServiceStub } from '../testing/notifications-service.stub';
-import { ItemRequestDataService } from '../../core/data/item-request-data.service';
-import { NotificationsService } from '../notifications/notifications.service';
-import { DSONameService } from '../../core/breadcrumbs/dso-name.service';
-import { DSONameServiceMock } from '../mocks/dso-name.service.mock';
-import { Item } from '../../core/shared/item.model';
-import { EPerson } from '../../core/eperson/models/eperson.model';
-import { ItemRequest } from '../../core/shared/item-request.model';
+import { NotificationsServiceStub } from '../../../shared/testing/notifications-service.stub';
+import { ItemRequestDataService } from '../../../core/data/item-request-data.service';
+import { NotificationsService } from '../../../shared/notifications/notifications.service';
+import { DSONameService } from '../../../core/breadcrumbs/dso-name.service';
+import { DSONameServiceMock } from '../../../shared/mocks/dso-name.service.mock';
+import { Item } from '../../../core/shared/item.model';
+import { EPerson } from '../../../core/eperson/models/eperson.model';
+import { ItemRequest } from '../../../core/shared/item-request.model';
 import { Location } from '@angular/common';
-import { BitstreamDataService } from '../../core/data/bitstream-data.service';
+import { BitstreamDataService } from '../../../core/data/bitstream-data.service';
 
 
 describe('BitstreamRequestACopyPageComponent', () => {
