+++ src/app/item-page/edit-item-page/edit-item-page.module.ts
@@ -36,6 +36,7 @@ import { ItemVersionHistoryComponent } from './item-version-history/item-version
 import { ItemAuthorizationsComponent } from './item-authorizations/item-authorizations.component';
 import { ObjectValuesPipe } from '../../shared/utils/object-values-pipe';
 import { ResourcePoliciesModule } from '../../shared/resource-policies/resource-policies.module';
+import { ItemVersionsModule } from '../versions/item-versions.module';
 
 
 /**
@@ -50,7 +51,8 @@ import { ResourcePoliciesModule } from '../../shared/resource-policies/resource-
     SearchPageModule,
     DragDropModule,
     ResourcePoliciesModule,
-    NgbModule
+    NgbModule,
+    ItemVersionsModule,
   ],
   declarations: [
     EditItemPageComponent,
