+++ src/app/item-page/item-shared.module.ts
@@ -7,10 +7,32 @@ import { TranslateModule } from '@ngx-translate/core';
 import { DYNAMIC_FORM_CONTROL_MAP_FN } from '@ng-dynamic-forms/core';
 import { dsDynamicFormControlMapFn } from '../shared/form/builder/ds-dynamic-form-ui/ds-dynamic-form-control-container.component';
 import { TabbedRelatedEntitiesSearchComponent } from './simple/related-entities/tabbed-related-entities-search/tabbed-related-entities-search.component';
+import { ItemVersionsDeleteModalComponent } from './versions/item-versions-delete-modal/item-versions-delete-modal.component';
+import { ItemVersionsSummaryModalComponent } from './versions/item-versions-summary-modal/item-versions-summary-modal.component';
+import { MetadataValuesComponent } from './field-components/metadata-values/metadata-values.component';
+import { DsoPageVersionButtonComponent } from '../shared/dso-page/dso-page-version-button/dso-page-version-button.component';
+import { PersonPageClaimButtonComponent } from '../shared/dso-page/person-page-claim-button/person-page-claim-button.component';
+import { GenericItemPageFieldComponent } from './simple/field-components/specific-field/generic/generic-item-page-field.component';
+import { MetadataRepresentationListComponent } from './simple/metadata-representation-list/metadata-representation-list.component';
+import { RelatedItemsComponent } from './simple/related-items/related-items-component';
+import { DsoPageOrcidButtonComponent } from '../shared/dso-page/dso-page-orcid-button/dso-page-orcid-button.component';
+
+const ENTRY_COMPONENTS = [
+  ItemVersionsDeleteModalComponent,
+  ItemVersionsSummaryModalComponent,
+];
 
 const COMPONENTS = [
+  ...ENTRY_COMPONENTS,
   RelatedEntitiesSearchComponent,
-  TabbedRelatedEntitiesSearchComponent
+  TabbedRelatedEntitiesSearchComponent,
+  MetadataValuesComponent,
+  DsoPageVersionButtonComponent,
+  PersonPageClaimButtonComponent,
+  GenericItemPageFieldComponent,
+  MetadataRepresentationListComponent,
+  RelatedItemsComponent,
+  DsoPageOrcidButtonComponent
 ];
 
 @NgModule({
@@ -30,7 +52,8 @@ const COMPONENTS = [
     {
       provide: DYNAMIC_FORM_CONTROL_MAP_FN,
       useValue: dsDynamicFormControlMapFn
-    }
+    },
+    ...ENTRY_COMPONENTS,
   ]
 })
 export class ItemSharedModule { }
