+++ src/app/item-page/simple/item-types/versioned-item/versioned-item.component.spec.ts
@@ -5,7 +5,7 @@ import { VersionHistoryDataService } from '../../../../core/data/version-history
 import { TranslateLoader, TranslateModule } from '@ngx-translate/core';
 import { VersionDataService } from '../../../../core/data/version-data.service';
 import { NotificationsService } from '../../../../shared/notifications/notifications.service';
-import { ItemVersionsSharedService } from '../../../../shared/item/item-versions/item-versions-shared.service';
+import { ItemVersionsSharedService } from '../../../versions/item-versions-shared.service';
 import { Item } from '../../../../core/shared/item.model';
 import { createSuccessfulRemoteDataObject$ } from '../../../../shared/remote-data.utils';
 import { buildPaginatedList } from '../../../../core/data/paginated-list.model';
@@ -20,6 +20,7 @@ import { ItemDataService } from '../../../../core/data/item-data.service';
 import { Version } from '../../../../core/shared/version.model';
 import { RouteService } from '../../../../core/services/route.service';
 import { TranslateLoaderMock } from '../../../../shared/testing/translate-loader.mock';
+import { ItemSharedModule } from '../../../item-shared.module';
 
 const mockItem: Item = Object.assign(new Item(), {
   bundles: createSuccessfulRemoteDataObject$(buildPaginatedList(new PageInfo(), [])),
@@ -66,6 +67,7 @@ describe('VersionedItemComponent', () => {
             useClass: TranslateLoaderMock,
           }
         }),
+        ItemSharedModule,
       ],
       providers: [
         { provide: VersionHistoryDataService, useValue: versionHistoryServiceSpy },
