+++ src/app/item-page/simple/item-types/versioned-item/versioned-item.component.ts
@@ -1,6 +1,6 @@
 import { Component } from '@angular/core';
 import { ItemComponent } from '../shared/item.component';
-import { ItemVersionsSummaryModalComponent } from '../../../../shared/item/item-versions/item-versions-summary-modal/item-versions-summary-modal.component';
+import { ItemVersionsSummaryModalComponent } from '../../../versions/item-versions-summary-modal/item-versions-summary-modal.component';
 import { getFirstCompletedRemoteData, getFirstSucceededRemoteDataPayload } from '../../../../core/shared/operators';
 import { RemoteData } from '../../../../core/data/remote-data';
 import { Version } from '../../../../core/shared/version.model';
@@ -9,7 +9,7 @@ import { NgbModal } from '@ng-bootstrap/ng-bootstrap';
 import { VersionHistoryDataService } from '../../../../core/data/version-history-data.service';
 import { TranslateService } from '@ngx-translate/core';
 import { VersionDataService } from '../../../../core/data/version-data.service';
-import { ItemVersionsSharedService } from '../../../../shared/item/item-versions/item-versions-shared.service';
+import { ItemVersionsSharedService } from '../../../versions/item-versions-shared.service';
 import { Router } from '@angular/router';
 import { WorkspaceitemDataService } from '../../../../core/submission/workspaceitem-data.service';
 import { SearchService } from '../../../../core/shared/search/search.service';
