+++ src/app/item-page/versions/item-versions-shared.service.spec.ts
@@ -2,15 +2,15 @@ import { TestBed } from '@angular/core/testing';
 
 import { ItemVersionsSharedService } from './item-versions-shared.service';
 import { ActivatedRoute } from '@angular/router';
-import { VersionDataService } from '../../../core/data/version-data.service';
-import { AuthService } from '../../../core/auth/auth.service';
-import { NotificationsService } from '../../notifications/notifications.service';
+import { VersionDataService } from '../../core/data/version-data.service';
+import { AuthService } from '../../core/auth/auth.service';
+import { NotificationsService } from '../../shared/notifications/notifications.service';
 import { TranslateService } from '@ngx-translate/core';
-import { VersionHistoryDataService } from '../../../core/data/version-history-data.service';
-import { WorkspaceitemDataService } from '../../../core/submission/workspaceitem-data.service';
-import { WorkflowItemDataService } from '../../../core/submission/workflowitem-data.service';
-import { createFailedRemoteDataObject, createSuccessfulRemoteDataObject } from '../../remote-data.utils';
-import { Version } from '../../../core/shared/version.model';
+import { VersionHistoryDataService } from '../../core/data/version-history-data.service';
+import { WorkspaceitemDataService } from '../../core/submission/workspaceitem-data.service';
+import { WorkflowItemDataService } from '../../core/submission/workflowitem-data.service';
+import { createFailedRemoteDataObject, createSuccessfulRemoteDataObject } from '../../shared/remote-data.utils';
+import { Version } from '../../core/shared/version.model';
 
 describe('ItemVersionsSharedService', () => {
   let service: ItemVersionsSharedService;
