+++ src/app/item-page/versions/item-versions-shared.service.ts
@@ -1,8 +1,8 @@
 import { Injectable } from '@angular/core';
-import { NotificationsService } from '../../notifications/notifications.service';
+import { NotificationsService } from '../../shared/notifications/notifications.service';
 import { TranslateService } from '@ngx-translate/core';
-import { RemoteData } from '../../../core/data/remote-data';
-import { Version } from '../../../core/shared/version.model';
+import { RemoteData } from '../../core/data/remote-data';
+import { Version } from '../../core/shared/version.model';
 
 @Injectable({
   providedIn: 'root'
