+++ src/app/item-page/versions/item-versions.component.scss
