+++ src/app/item-page/versions/notice/item-versions-notice.component.spec.ts
@@ -3,13 +3,13 @@ import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 import { TranslateModule } from '@ngx-translate/core';
 import { RouterTestingModule } from '@angular/router/testing';
 import { NO_ERRORS_SCHEMA } from '@angular/core';
-import { VersionHistory } from '../../../../core/shared/version-history.model';
-import { Version } from '../../../../core/shared/version.model';
-import { Item } from '../../../../core/shared/item.model';
-import { VersionHistoryDataService } from '../../../../core/data/version-history-data.service';
+import { VersionHistory } from '../../../core/shared/version-history.model';
+import { Version } from '../../../core/shared/version.model';
+import { Item } from '../../../core/shared/item.model';
+import { VersionHistoryDataService } from '../../../core/data/version-history-data.service';
 import { By } from '@angular/platform-browser';
-import { createSuccessfulRemoteDataObject$ } from '../../../remote-data.utils';
-import { createPaginatedList } from '../../../testing/utils.test';
+import { createSuccessfulRemoteDataObject$ } from '../../../shared/remote-data.utils';
+import { createPaginatedList } from '../../../shared/testing/utils.test';
 import { of } from 'rxjs';
 import { take } from 'rxjs/operators';
